Module rust_eterm::terms

Currently implemented term types (into binary, String):

INTEGER_EXT, SMALL_INTEGER_EXT (for {i,u}{8,16,32,64,128,size})

FLOAT_EXT (for f32, f64)

NIL_EXT (for ENil)

LIST_EXT (for EList, ENonProperList)

EXPORT_EXT (for EExport)

LARGE_TUPLE_EXT, SMALL_TUPLE_EXT (for ETuple)

MAP_EXT (for EMap)

BINARY_EXT (for EBinary)

SMALL_BIG_EXT (for BigInt, BigUint, u32, i64, u64, i128, u128)

LARGE_BIG_EXT (for BigInt, BigUint)

NEW_PORT_EXT (for EPort)

NEW_PID_EXT (for EPid)

ATOM_UTF8_EXT, SMALL_ATOM_UTF8_EXT (for EAtom)

STRING_EXT (for EString)

Currently implemented term types (from binary, String)

None

Currently implemented term types (both from binary, String and into binary, String)

None

Not yet implemented term/value types:

NEWER_REFERENCE_EXT

BIT_BINARY_EXT

ATOM_CACHE_REF (distribution header will probably not be supported, or not soon at least)

FUN_EXT (seems unnecessary for now, maybe in the future?)

NEW_FUN_EXT (seems unnecessary for now, maybe in the future?)

REFERENCE_EXT (deprecated, decoding support will be added)

NEW_REFERENCE_EXT (decoding support will be added)

ATOM_EXT (deprecated, decoding support will be added)

SMALL_ATOM_EXT (deprecated, decoding support will be added)

PORT_EXT (decoding support will be added)

PID_EXT (decoding support will be added)

ETF_VERSION

DIST_HDR_NORMAL

DIST_HDR_FRAG_START

DIST_HDR_FRAG_CONT

DIST_HDR_COMPRESSED

Modules

decode

Structs

EAtom
Describes an ATOM_UTF8_EXT term and a SMALL_ATOM_UTF8_EXT term.

EExport
Describes an Erlang EXPORT_EXT term.

EList
Represents a proper LIST_EXT term with a nil tail.

EMap
Describes an Erlang Map

ENil
Represents an Erlang NIL_EXT term.

ENonProperList
Describes a LIST_EXT term with a possible non-nil tail.

EPid
Describes an Erlang PID.

EPort
Describes an Erlang Port

EString
Describes an STRING_EXT term.

ETuple
Represents a LARGE_TUPLE_EXT or a SMALL_TUPLE_EXT term.

Enums

DistHeaderTag

TermTag

Constants

ETF_VERSION
This is the code of the start of a message. This message will contain an encoded tag or a distribution header.

Traits

ETerm
A type that can be converted to an Erlang Binary Term format and two valid Erlang String Term representations.