Module rust_eterm::terms

Partially implemented term types (into binary, String):

ATOM_UTF8_EXT, SMALL_ATOM_UTF8_EXT (for EAtom, quoted atoms are not yet supported)

STRING_EXT (for EString, Strings that require escaping are not yet supported)

Currently implemented term types (into binary, String):

INTEGER_EXT, SMALL_INTEGER_EXT (for {i,u}{8,16,32,64,128,size})

FLOAT_EXT (for f32, f64)

NIL_EXT (for ENil)

LIST_EXT (for EList, ENonProperList)

EXPORT_EXT (for EExport)

LARGE_TUPLE_EXT, SMALL_TUPLE_EXT (for ETuple)

MAP_EXT (for EMap)

BINARY_EXT (for EBinary)

SMALL_BIG_EXT (for BigInt, BigUint, u32, i64, u64, i128, u128)

LARGE_BIG_EXT (for BigInt, BigUint)

NEW_PORT_EXT (for EPort)

NEW_PID_EXT (for EPid)

Currently implemented term types (from binary, String)

None

Currently implemented term types (both from binary, String and into binary, String)

None

Not yet implemented term types:

NEWER_REFERENCE_EXT

BIT_BINARY_EXT

No implementation plans

ATOM_CACHE_REF (distribution header will probably not be supported, or not soon at least)

FUN_EXT (seems unnecessary for now, maybe in the future?)

NEW_FUN_EXT (seems unnecessary for now, maybe in the future?)

REFERENCE_EXT (deprecated, decoding support will be added)

NEW_REFERENCE_EXT (decoding support will be added)

ATOM_EXT (deprecated, decoding support will be added)

SMALL_ATOM_EXT (deprecated, decoding support will be added)

PORT_EXT (decoding support will be added)

PID_EXT (decoding support will be added)

Structs

EAtom
Describes an ATOM_UTF8_EXT term and a SMALL_ATOM_UTF8_EXT term.

EExport
Describes an Erlang EXPORT_EXT term.

EList
Represents a proper LIST_EXT term with a nil tail.

EMap
Describes an Erlang Map

ENil
Represents an Erlang NIL_EXT term.

ENonProperList
Describes a LIST_EXT term with a possible non-nil tail.

EPid
Describes an Erlang PID.

EPort
Describes an Erlang Port

EString
Describes an ATOM_UTF8_EXT term and a SMALL_ATOM_UTF8_EXT term.

ETermBinary
Binary representation for an ETerm.

ETermPrettyString
Pretty String representation for an Eterm.

ETermString
Ugly String representation for an ETerm.

ETuple
Represents a LARGE_TUPLE_EXT or a SMALL_TUPLE_EXT term with a nil tail.

Statics

ATOM_CACHE_REF
Refers to the atom with AtomCacheReferenceIndex in the distribution header.

ATOM_EXT

ATOM_UTF8_EXT

BINARY_EXT
Binaries are byte-arrays. They are represented as bitstrings or binaries in the Erlang language.

BIT_BINARY_EXT
This term represents a bitstring whose length in bits does not have to be a multiple of 8.

DIST_HDR

DIST_HDR_COMPRESSED

DIST_HDR_FRAGMENT

DIST_HDR_FRAGMENTED

DIST_HDR_NORMAL

EXPORT_EXT
Encodes functions of the shape fun M:F/A.

FLOAT_EXT
A finite float (i.e. not inf, -inf or NaN) is stored in string format.

FUN_EXT
Old encoding of internal functions: fun F/A and fun(Arg1,..) -> ... end.

INTEGER_EXT
Signed 32-bit integer.

LARGE_BIG_EXT
Integer representation of an integer N where -2^(2^32) < N < 2^(2^32).

LARGE_TUPLE_EXT
Same as SMALL_TUPLE_EXT except that Arity is an unsigned 4 byte integer.

LIST_EXT
The representation for a non-empty list.

MAP_EXT
Encodes a map.

NEWER_REFERENCE_EXT
Encodes a reference term.

NEW_FLOAT_EXT

NEW_FUN_EXT
Encoding of internal functions: fun F/A and fun(Arg1,..) -> ... end.

NEW_PID_EXT
Encodes an Erlang process identifier object.

NEW_PORT_EXT
Encodes a port identifier (obtained from erlang:open_port/2).

NEW_REFERENCE_EXT
Deprecated method for encoding a reference term.

NIL_EXT
The representation for an empty list, that is, the Erlang syntax [].

PID_EXT
Same as NEW_PID_EXT except the Creation field is only one byte and only two bits are significant, the rest are to be 0.

PORT_EXT
Same as NEW_PORT_EXT except the Creation field is only one byte and only two bits are significant, the rest are to be 0.

REFERENCE_EXT
Deprecated method for encoding a reference term.

SMALL_ATOM_EXT

SMALL_ATOM_UTF8_EXT

SMALL_BIG_EXT
Integer representation of an integer N where -2^256 < N < 2^256.

SMALL_INTEGER_EXT
Unsigned 8-bit integer.

SMALL_TUPLE_EXT
Encodes a tuple.

STRING_EXT
String does not have a corresponding Erlang representation, but is an optimization for sending lists of bytes (integer in the range 0-255) more efficiently over the distribution.

Traits

ETerm
A type that can be converted to an Erlang Binary Term format and two valid Erlang String Term representations.

To
Replacement for std::convert::Into<T> that doesn't require Sized.

TryTo
Replacement for std::convert::TryInto<T> that doesn't require Sized.